+++ Documentation/faq/releases.rst
@@ -238,7 +238,7 @@ Q: Are all the DPDK releases that OVS versions work with maintained?
     The latest information about DPDK stable and LTS releases can be found
     at `DPDK stable`_.
 
-.. _DPDK stable: http://doc.dpdk.org/guides-23.11/contributing/stable.html
+.. _DPDK stable: http://doc.dpdk.org/guides-24.11/contributing/stable.html
 
 Q: I get an error like this when I configure Open vSwitch: